@@ -11,35 +11,36 @@
     "require": {
         "awcodes/shout": "^2.0.2",
         "barryvdh/laravel-dompdf": "^2.0",
-        "cartalyst/converter": "^8.0",
+        "dompdf/dompdf": "^2.0.7",
+        "cartalyst/converter": "^8.0|^9.0",
         "doctrine/dbal": "^3.6",
         "ext-bcmath": "*",
         "ext-exif": "*",
         "ext-intl": "*",
         "filament/filament": "^3.2.25",
-        "filament/spatie-laravel-media-library-plugin": "^3.0-stable",
+        "filament/spatie-laravel-media-library-plugin": "^3.2",
         "guzzlehttp/guzzle": "^7.3",
-        "illuminate/http": "^10.0",
-        "kalnoy/nestedset": "^6.0",
-        "laravel/framework": "^10.0",
+        "kalnoy/nestedset": "^v6.x-dev",
+        "laravel/framework": "^10.0|^11.0",
         "laravel/scout": "^10.0",
-        "leandrocfe/filament-apex-charts": "^3.1.2",
+        "leandrocfe/filament-apex-charts": "^3.1.3",
         "livewire/livewire": "^3.0",
         "lukascivil/treewalker": "0.9.1",
-        "marvinosswald/filament-input-select-affix": "^0.1.0",
-        "php": "^8.1",
+        "marvinosswald/filament-input-select-affix": "^0.2.0",
+        "meilisearch/meilisearch-php": "^v1.6.0",
+        "php": "^8.2",
         "spatie/laravel-activitylog": "^4.4",
-        "spatie/laravel-blink": "^1.6",
-        "spatie/laravel-medialibrary": "^10.0.0",
-        "spatie/laravel-permission": "^5.10",
+        "spatie/laravel-blink": "^1.7",
+        "spatie/laravel-medialibrary": "^11.0.0",
+        "spatie/laravel-permission": "^6.4",
         "stripe/stripe-php": "^7.114",
-        "technikermathe/blade-lucide-icons": "^v2.24.0"
+        "technikermathe/blade-lucide-icons": "^v3.0"
     },
     "require-dev": {
+        "larastan/larastan": "^2.9",
         "laravel/pint": "1.13.1",
         "mockery/mockery": "^1.4.4",
-        "larastan/larastan": "^2.9",
-        "orchestra/testbench": "^8.0",
+        "orchestra/testbench": "^8.0|^9.0",
         "pestphp/pest": "^2.0",
         "pestphp/pest-plugin-laravel": "^2.0",
         "symplify/monorepo-builder": "^10.0"
