@@ -11,7 +11,7 @@
     }
   ],
   "require": {
-      "php": "^8.1",
+      "php": "^8.2",
       "lunarphp/core": "self.version",
       "stripe/stripe-php": "^7.114"
   },
