@@ -11,6 +11,7 @@ use Lunar\Admin\Support\FieldTypes\Number;
 use Lunar\Admin\Support\FieldTypes\TextField;
 use Lunar\Admin\Support\FieldTypes\Toggle;
 use Lunar\Admin\Support\FieldTypes\TranslatedText;
+use Lunar\Admin\Support\FieldTypes\Vimeo;
 use Lunar\Admin\Support\FieldTypes\YouTube;
 use Lunar\FieldTypes\Dropdown as DrodownFieldType;
 use Lunar\FieldTypes\File as FileFieldType;
@@ -19,6 +20,7 @@ use Lunar\FieldTypes\Number as NumberFieldType;
 use Lunar\FieldTypes\Text as TextFieldType;
 use Lunar\FieldTypes\Toggle as ToggleFieldType;
 use Lunar\FieldTypes\TranslatedText as TranslatedTextFieldType;
+use Lunar\FieldTypes\Vimeo as VimeoFieldType;
 use Lunar\FieldTypes\YouTube as YouTubeFieldType;
 use Lunar\Models\Attribute;
 
@@ -31,6 +33,7 @@ class AttributeData
         TranslatedTextFieldType::class => TranslatedText::class,
         ToggleFieldType::class => Toggle::class,
         YouTubeFieldType::class => YouTube::class,
+        VimeoFieldType::class => Vimeo::class,
         NumberFieldType::class => Number::class,
         FileFieldType::class => File::class,
     ];
@@ -44,13 +47,7 @@ class AttributeData
         return $fieldType::getFilamentComponent($attribute)->label(
             $attribute->translate('name')
         )
-            ->formatStateUsing(fn ($state) => ($state ?: (new $attribute->type))->getValue())
-            ->dehydrateStateUsing(function ($state) use ($attribute) {
-                $field = new $attribute->type;
-                $field->setValue($state);
-
-                return $field;
-            })
+            ->formatStateUsing(fn ($state) => ($state ?: new $attribute->type))
             ->required($attribute->required)
             ->default($attribute->default_value);
     }
