@@ -52,7 +52,8 @@ export default defineConfig({
                 items: [
                     {text: 'Changelog', link: '/core/upgrading'},
                     {text: 'Contributing', link: '/core/contributing'},
-                    {text: 'Roadmap', link: 'https://github.com/orgs/lunarphp/projects/8'}
+                    {text: 'Roadmap', link: 'https://github.com/orgs/lunarphp/projects/8'},
+                    {text: '0.x Docs', link: 'https://v0.lunarphp.io/'},
                 ]
             }
         ],
