@@ -8,6 +8,7 @@ use Filament\Tables;
 use Filament\Tables\Table;
 use Illuminate\Contracts\Support\Htmlable;
 use Illuminate\Database\Eloquent\Model;
+use Lunar\Admin\Events\ChildCollectionCreated;
 use Lunar\Admin\Filament\Resources\CollectionResource;
 use Lunar\Admin\Support\Pages\BaseManageRelatedRecords;
 use Lunar\Admin\Support\Tables\Actions\Collections\CreateChildCollection;
@@ -73,7 +74,9 @@ class ManageCollectionChildren extends BaseManageRelatedRecords
                 return CollectionResource::getUrl('edit', ['record' => $record]);
             }),
         ])->headerActions([
-            CreateChildCollection::make('createChildCollection'),
+            CreateChildCollection::make('createChildCollection')->after(
+                fn () => ChildCollectionCreated::dispatch($this->getRecord())
+            ),
         ]);
     }
 }
