@@ -59,7 +59,7 @@ class ProductLimitationRelationManager extends RelationManager
                 }),
             ])->columns([
                 Tables\Columns\SpatieMediaLibraryImageColumn::make('purchasable.thumbnail')
-                    ->collection('images')
+                    ->collection(config('lunar.media.collection'))
                     ->conversion('small')
                     ->limit(1)
                     ->square()
