@@ -10,6 +10,7 @@ use Filament\Tables\Table;
 use Illuminate\Contracts\Support\Htmlable;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Validation\Rules\Unique;
+use Lunar\Admin\Events\ProductPricingUpdated;
 use Lunar\Facades\DB;
 use Lunar\Models\Currency;
 use Lunar\Models\CustomerGroup;
@@ -148,8 +149,10 @@ class CustomerGroupPricingRelationManager extends RelationManager
                     return $data;
                 })->label(
                     __('lunarpanel::relationmanagers.customer_group_pricing.table.actions.create.label')
-                )
-                    ->modalHeading(__('lunarpanel::relationmanagers.customer_group_pricing.table.actions.create.modal.heading')),
+                )->modalHeading(__('lunarpanel::relationmanagers.customer_group_pricing.table.actions.create.modal.heading'))
+                    ->after(
+                        fn () => ProductPricingUpdated::dispatch($this->getOwnerRecord())
+                    ),
             ])
             ->actions([
                 Tables\Actions\EditAction::make()->mutateFormDataUsing(function (array $data): array {
@@ -159,7 +162,9 @@ class CustomerGroupPricingRelationManager extends RelationManager
                     $data['price'] = (int) ($data['price'] * $currencyModel->factor);
 
                     return $data;
-                }),
+                })->after(
+                    fn () => ProductPricingUpdated::dispatch($this->getOwnerRecord())
+                ),
                 Tables\Actions\DeleteAction::make(),
             ]);
     }
