@@ -195,7 +195,7 @@ class StaffResource extends BaseResource
             ]);
     }
 
-    public static function getRelations(): array
+    public static function getDefaultRelations(): array
     {
         return [
             //
