@@ -3,9 +3,11 @@
 namespace Lunar\Admin\Support\Actions\Collections;
 
 use Filament\Actions\CreateAction;
+use Filament\Forms\Components\TextInput;
 use Illuminate\Database\Eloquent\Model;
 use Lunar\Admin\Support\Actions\Traits\CreatesChildCollections;
 use Lunar\Admin\Support\Forms\Components\TranslatedText;
+use Lunar\Models\Attribute;
 use Lunar\Models\Collection;
 
 class CreateChildCollection extends CreateAction
@@ -24,8 +26,17 @@ class CreateChildCollection extends CreateAction
             $this->success();
         });
 
+        $attribute = Attribute::where('attribute_type', '=', Collection::class)
+            ->where('handle', '=', 'name')->first();
+
+        $formInput = TextInput::class;
+
+        if ($attribute?->type == \Lunar\FieldTypes\TranslatedText::class) {
+            $formInput = TranslatedText::class;
+        }
+
         $this->form([
-            TranslatedText::make('name')->required(),
+            $formInput::make('name')->required(),
         ]);
 
         $this->label(
