@@ -42,19 +42,19 @@ trait UpdatesOrderStatus
                 }
 
                 return ! count($get('mailers') ?: [])
-                    || ! ($record?->billingAddress?->contact_email && $record->shippingAddress->contact_email);
+                    || ! ($record?->billingAddress?->contact_email && $record->shippingAddress?->contact_email);
             })->afterStateHydrated(function (Order $record = null, Forms\Components\CheckboxList $component) {
                 $emails = collect([
-                    $record?->billingAddress->contact_email,
-                    $record?->shippingAddress->contact_email,
+                    $record?->billingAddress?->contact_email,
+                    $record?->shippingAddress?->contact_email,
                 ])->filter()->unique()->map(
                     fn ($email) => $email
                 )->toArray();
                 $component->state($emails);
             })->options(function (Order $record = null) {
                 return collect([
-                    $record?->billingAddress->contact_email,
-                    $record?->shippingAddress->contact_email,
+                    $record?->billingAddress?->contact_email,
+                    $record?->shippingAddress?->contact_email,
                 ])->filter()->unique()->mapWithKeys(
                     fn ($email) => [$email => $email]
                 )->toArray();
