@@ -52,7 +52,7 @@ class ShippingExclusionRelationManager extends RelationManager
         return $table
             ->columns([
                 Tables\Columns\SpatieMediaLibraryImageColumn::make('purchasable.thumbnail')
-                    ->collection('images')
+                    ->collection(config('lunar.media.collection'))
                     ->conversion('small')
                     ->limit(1)
                     ->square()
