@@ -3,8 +3,8 @@
 namespace Lunar\Shipping\Models;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
+use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Lunar\Base\BaseModel;
-use Lunar\Shipping\Facades\Shipping;
 use Lunar\Shipping\Factories\ShippingZonePostcodeFactory;
 
 class ShippingZonePostcode extends BaseModel
@@ -31,21 +31,16 @@ class ShippingZonePostcode extends BaseModel
 
     /**
      * Return the shipping zone relationship.
-     *
-     * @return \Illuminate\Database\Eloquent\Relations\BelongsTo
      */
-    public function shippingZone()
+    public function shippingZone(): BelongsTo
     {
         return $this->belongsTo(ShippingZone::class);
     }
 
     /**
      * Setter for postcode attribute.
-     *
-     * @param  string  $value
-     * @return void
      */
-    public function setPostcodeAttribute($value)
+    public function setPostcodeAttribute(string $value): void
     {
         $this->attributes['postcode'] = str_replace(' ', '', $value);
     }
