@@ -8,23 +8,23 @@ use Lunar\Shipping\Facades\Shipping;
 
 class OrderObserver
 {
-    public function updated(Order $order)
+    public function updated(Order $order): void
     {
         $this->updateShippingZone(
             $order
         );
     }
 
-    public function created(Order $order)
+    public function created(Order $order): void
     {
         $this->updateShippingZone(
             $order
         );
     }
 
-    protected function updateShippingZone(Order $order)
+    protected function updateShippingZone(Order $order): void
     {
-        $shippingAddress = $order->shippingAddress ?: $order->cart->shippingAddress;
+        $shippingAddress = $order->shippingAddress ?: $order->cart?->shippingAddress;
 
         if ($shippingAddress && $shippingAddress->postcode) {
             $postcodeLookup = new PostcodeLookup(
@@ -49,7 +49,7 @@ class OrderObserver
     /**
      * Called when we're about to index the order.
      **/
-    public function indexing(Order $order)
+    public function indexing(Order $order): void
     {
         $order->addSearchableAttribute('shipping_zone', $order->meta?->shipping_zone ?? null);
     }
