@@ -18,22 +18,16 @@ class ShippingRateResolver
 
     /**
      * The country to use when resolving.
-     *
-     * @var Country
      */
     protected ?Country $country = null;
 
     /**
      * The state to use when resolving.
-     *
-     * @var State
      */
     protected ?string $state = null;
 
     /**
      * The postcode to use when resolving.
-     *
-     * @var string
      */
     protected ?string $postcode = null;
 
@@ -46,8 +40,6 @@ class ShippingRateResolver
 
     /**
      * Initialise the resolver.
-     *
-     * @param  Cart  $cart
      */
     public function __construct(Cart $cart = null)
     {
@@ -138,15 +130,15 @@ class ShippingRateResolver
             State::whereName($this->state)->first()
         )->postcode(
             new PostcodeLookup(
-                postcode: $this->postcode,
-                country: $this->country
+                country: $this->country,
+                postcode: $this->postcode
             )
         )->get();
 
         $shippingRates = collect();
 
         foreach ($zones as $zone) {
-            $shippingRates = $zone->rates
+            $zoneShippingRates = $zone->rates
                 ->reject(function ($rate) {
                     $method = $rate->shippingMethod;
 
@@ -169,7 +161,7 @@ class ShippingRateResolver
                     return true;
                 });
 
-            foreach ($shippingRates as $shippingRate) {
+            foreach ($zoneShippingRates as $shippingRate) {
                 $shippingRates->push(
                     $shippingRate
                 );
