@@ -6,11 +6,11 @@ use Illuminate\Support\ServiceProvider;
 use Lunar\Base\ShippingModifiers;
 use Lunar\Models\Order;
 use Lunar\Models\Product;
-use Lunar\Observers\OrderObserver;
 use Lunar\Shipping\Interfaces\ShippingMethodManagerInterface;
 use Lunar\Shipping\Managers\ShippingManager;
 use Lunar\Shipping\Models\ShippingExclusion;
 use Lunar\Shipping\Models\ShippingZone;
+use Lunar\Shipping\Observers\OrderObserver;
 
 class ShippingServiceProvider extends ServiceProvider
 {
