@@ -14,14 +14,14 @@ class TestMediaDefinition implements \Lunar\Base\MediaDefinitionsInterface
 
     public function registerMediaCollections(HasMedia $model): void
     {
-        $model->addMediaCollection('images');
+        $model->addMediaCollection(config('lunar.media.collection'));
         $model->addMediaCollection('videos');
     }
 
     public function getMediaCollectionTitles(): array
     {
         return [
-            'images' => 'Images',
+            config('lunar.media.collection') => 'Images',
             'videos' => 'Videos',
         ];
     }
@@ -29,7 +29,7 @@ class TestMediaDefinition implements \Lunar\Base\MediaDefinitionsInterface
     public function getMediaCollectionDescriptions(): array
     {
         return [
-            'images' => 'Images',
+            config('lunar.media.collection') => 'Images',
             'videos' => 'Videos',
         ];
     }
